-Fixed and enabled victory sitreps by making the part of the stringtable entry that contains the empire name reference in the top-level string, and adding a new TEXT_TAG to VarText that is replaced by the indicated text.  This is a bit awkward since it uses a slightly different syntax from the rest of the text substitution system, but this happens in the midst of VarText substitution code that can't easily be switched to use standard boost::format to do the substitution.  The two existing victory conditions are now formatted to be appended to the common root text of the SITREP_VICTORY string.  This might be awkward for some translations, but it's the only reasonably simple solution I could work out. -Replaced the ClientUI g_string_id_lut array with SitRepEntry::SitRepTemplateString that returns the appropriate string template for each sitrep EntryType -Added TemplateString member function to SitRepEntry that looks up the sitrep's entry type and returns the string, avoiding the need for external use of a look up table at all -Added a constructor to SitRepEntry that takes an EntryType, and remove the SetType function which is no longer needed -Used VarText static tag strings when registering font tags and when scanning strings for tags in TextLinker::FindLinks to ensure they're all consistent and only defined in one place -Added new VarText tags (and linking support) for building types (as distinct from building objects), specials, ship hulls and ship parts. -Minor tweak to quiet type conversion warning to parameter of ProductionInfoPanel -Made GetSpecial return a const Special as there was no need to return a non-const Special, as Specials shouldn't be modified after they're loaded. -Misc. grooming
svn path=/trunk/FreeOrion/; revision=3595

+++ UI/CUIControls.cpp
@@ -1219,7 +1219,7 @@ const int ProductionInfoPanel::CORNER_RADIUS = 9;
 const GG::Y ProductionInfoPanel::VERTICAL_SECTION_GAP(4);
 
 ProductionInfoPanel::ProductionInfoPanel(GG::X w, GG::Y h, const std::string& title, const std::string& points_str,
-                                         double border_thickness, const GG::Clr& color, const GG::Clr& text_and_border_color) :
+                                         float border_thickness, const GG::Clr& color, const GG::Clr& text_and_border_color) :
     GG::Wnd(GG::X0, GG::Y0, w, h, GG::Flags<GG::WndFlag>()),
     m_border_thickness(border_thickness),
     m_color(color),
